Fix student submission path: image content type, attemptNumber encoding
- Uploads had no content type, so storage.rules (image/* only) rejected
  every drawing and exercise photo. Set it from the file extension.
- The Apple Functions SDK wraps a Swift Int as an Int64Value object;
  correct_submission rejected it as an invalid attemptNumber. Send Int32
  and accept the wrapped form server-side.
- CLAUDE_PROVIDER=fake: canned Claude answers, emulator only, so the
  whole flow can be clicked through without Vertex AI or an API key.

+++ Core/Services/CorrectionService.swift
@@ -61,7 +61,9 @@ final class CorrectionService {
             "expectedAnswer": expectedAnswer,
             "statement": statement,
             "submissionID": submissionID,
-            "attemptNumber": attemptNumber,
+            // Int32: the Functions SDK wraps a 64-bit Int as an Int64Value
+            // object, which the Python callable receives as a dict.
+            "attemptNumber": Int32(attemptNumber),
             "notationStrict": notationStrict
         ]
